chore: move admin methods to different file
